+++ src/Core/GraphDimension.cpp
@@ -8,9 +8,9 @@ namespace Pathfinding::Core
         currentNumberOfNodes_ = numOfNodes;
     }
 
-    int32_t GraphDimension::currentNumberOfNodesIndex() const
+    int64_t GraphDimension::currentNumberOfNodesIndex() const
     {
-        return static_cast<int32_t>(currentNumberOfNodes_);
+        return static_cast<int64_t>(currentNumberOfNodes_);
     }
 
     NumberOfNodes GraphDimension::currentNumberOfNodes() const
@@ -18,14 +18,14 @@ namespace Pathfinding::Core
         return currentNumberOfNodes_;
     }
 
-    int32_t GraphDimension::currentNodeSideLength() const
+    int64_t GraphDimension::currentNodeSideLength() const
     {
-        return NODE_SIDE_LENGTH[static_cast<int32_t>(currentNumberOfNodes_)];
+        return NODE_SIDE_LENGTH[static_cast<int64_t>(currentNumberOfNodes_)];
     }
 
-    int32_t GraphDimension::currentNumberOfNodesInRow() const
+    int64_t GraphDimension::currentNumberOfNodesInRow() const
     {
-        return NUMBER_OF_NODES_IN_ROW[static_cast<int32_t>(currentNumberOfNodes_)];
+        return NUMBER_OF_NODES_IN_ROW[static_cast<int64_t>(currentNumberOfNodes_)];
     }
 
     bool GraphDimension::canShowNodeInfo() const
@@ -33,17 +33,17 @@ namespace Pathfinding::Core
         return currentNumberOfNodes_ == NUMBER_OF_NODES_FOR_RENDER_INFO;
     }
 
-    void GraphDimension::setCurrentNumberOfNodesIndex(int32_t index)
+    void GraphDimension::setCurrentNumberOfNodesIndex(int64_t index)
     {
         currentNumberOfNodes_ = static_cast<NumberOfNodes>(index);
     }
 
-    int32_t GraphDimension::width() const
+    int64_t GraphDimension::width() const
     {
         return currentNumberOfNodesInRow();
     }
     
-    int32_t GraphDimension::height() const
+    int64_t GraphDimension::height() const
     {
         return currentNumberOfNodesInRow();
     }
